Remove curses library
I have decided to remove the curses library to supply the desired which is "getch()" function to pause system before ending the application. I simply couldn't link it. This is something new I will have to learn at some point... but, for now, I will keep the application an a while loop and give the user a chance to read the output and volitionally end the application.